workflows/crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "A web crawler of arxiv and tech blog sites."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "loguru~=0.7.3",
  "httpx~=0.28.1",
  "feedparser~=6.0.12",
  "certifi~=2025.11.12",
  "truststore~=0.10.4",
  "pydantic~=2.12.5",
  "tenacity~=9.1.2",
  "defusedxml~=0.7.1",
  "aiolimiter~=1.2.1",
  "google-cloud-storage~=3.9.0"
]

[dependency-groups]
dev = []
test = [
  "pytest~=9.0.1",
  "pytest-asyncio>=1.3.0",
  "pytest-console-scripts~=1.4.1",
  "pytest-mock>=3.15.1",
  "pytest-xdist>=3.8.0",
]
lint = [
  "ruff~=0.14.7",
  "mypy~=1.19.0",
  "types-defusedxml~=0.7.0.20250822",
  "ty~=0.0.18",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.mypy]
python_version = "3.13"
exclude = [".venv"]
# plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true

# [tool.pydantic-mypy]
# init_forbid_extra = true
# init_typed = true
# warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["google", "google.*", "google.cloud", "google.cloud.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto" # マーカーを省略しても自動で非同期テストを認識する
